fix: generic send max tests
